move events under plugin since they're only used by plugins
