--- pizzaserver-server/src/main/java/io/github/willqi/pizzaserver/server/events/Cancellable.java
@@ -1,4 +1,0 @@
-package io.github.willqi.pizzaserver.server.events;
-
-public interface Cancellable {
-}
