+++ pizzaserver-server/src/main/java/io/github/willqi/pizzaserver/server/network/handlers/LoginPacketHandler.java
@@ -3,7 +3,7 @@ package io.github.willqi.pizzaserver.server.network.handlers;
 import io.github.willqi.pizzaserver.server.Server;
 import io.github.willqi.pizzaserver.commons.server.Difficulty;
 import io.github.willqi.pizzaserver.server.data.ServerOrigin;
-import io.github.willqi.pizzaserver.server.events.player.PreLoginEvent;
+import io.github.willqi.pizzaserver.server.plugin.events.player.PreLoginEvent;
 import io.github.willqi.pizzaserver.server.network.BedrockClientSession;
 import io.github.willqi.pizzaserver.server.network.BedrockPacketHandler;
 import io.github.willqi.pizzaserver.server.network.protocol.ServerProtocol;
