+++ pizzaserver-server/src/main/java/io/github/willqi/pizzaserver/server/plugin/events/Cancellable.java
@@ -1,0 +1,4 @@
+package io.github.willqi.pizzaserver.server.plugin.events;
+
+public interface Cancellable {
+}
