+++ pizzaserver-server/src/main/java/io/github/willqi/pizzaserver/server/plugin/events/player/PreLoginEvent.java
@@ -1,7 +1,7 @@
-package io.github.willqi.pizzaserver.server.events.player;
+package io.github.willqi.pizzaserver.server.plugin.events.player;
 
 import io.github.willqi.pizzaserver.server.player.Player;
-import io.github.willqi.pizzaserver.server.events.generic.PlayerEvent;
+import io.github.willqi.pizzaserver.server.plugin.events.generic.PlayerEvent;
 
 public class PreLoginEvent extends PlayerEvent {
